Refactor multisig and netting types to use int32 for thresholds and statuses
Updated various types and keeper logic in multisig and netting modules to use int32 for threshold and status fields for consistency and protobuf compatibility. Adjusted related test code, property-based tests, and helper functions to match the new types. Also improved test setup and mock implementations for better reliability and maintainability.

+++ cosmos/testutil/setup.go
@@ -172,8 +172,8 @@ func GenValidatorSet() gopter.Gen {
 		gen.UInt64(),
 	).Map(func(values []interface{}) types.ValidatorSet {
 		validators := values[0].([]types.Validator)
-		threshold := len(validators)*2/3 + 1 // 2/3 + 1 threshold
-		
+		threshold := int32(len(validators)*2/3 + 1) // 2/3 + 1 threshold
+
 		return types.ValidatorSet{
 			Validators:   validators,
 			Threshold:    threshold,
@@ -199,7 +199,7 @@ func GenMintCommand() gopter.Gen {
 			Amount:      values[3].(math.Int),
 			Signatures:  []types.ECDSASignature{}, // Empty initially
 			CreatedAt:   values[4].(int64),
-			Status:      types.CommandStatusPending,
+			Status:      int32(types.CommandStatusPending),
 		}
 	})
 }
